petzi-app/requirements.txt
bidict==0.22.1
blinker==1.7.0
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
configparser==6.0.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
gevent==23.9.1
greenlet==3.0.1
gunicorn==20.1.0
h11==0.14.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pika==1.3.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
python-engineio==4.8.0
python-socketio==5.10.0
requests==2.31.0
setuptools==69.0.3
simple-websocket==1.0.0
SQLAlchemy==2.0.23
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
wsproto==1.2.0
zope.event==5.0
zope.interface==6.1
prometheus-client==0.19.0
Faker==20.1.0
